% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gen_tmle.R
\name{gentmle}
\alias{gentmle}
\title{gen_tmle}
\usage{
gentmle(initdata, estimate_fun, update_fun, max_iter = 100, ...)
}
\arguments{
\item{update_fun,}{Function for update step}

\item{max_iter,}{Maximum number of iteration steps}

\item{...,}{Extra arguments that can be passed to update_fun and estimate_fun}

\item{estiamte_fun}{Function for estimation step}
}
\description{
General TMLE function that takes care of the bookkeeping of estimation and update steps.
}

